refactor: optimize token value loss threshold calculation
