+++ packages/widget/src/pages/SwapPage/index.ts
@@ -1,2 +1,1 @@
 export * from './SwapPage';
-export * from './utils';
